Added ‘Notes’ to ‘Job’ so users can record some notes about each job application.
Adding a new job application to the tracker:
	- tapping on ‘+’ in HomeViewController will present a sheet (AddNewJobViewController).
	- users can enter text for the company, position, location, choose a status from the popup button, and add notes.

Editing a previously saved job application:
	- users can edit a job application by tapping on its cell.
	- the job application details will be presented in a sheet (JobDetailViewController).
	- users can make changes and tap ‘Done’ to save them or ‘Cancel’ to discard them.

Added a status label to both sheets to indicate that the status can be chosen from the popup button beside it.

The dashboard now shows dynamic data and updates when new applications are added or when existing ones are edited.

Implemented CoreData for persistence. If there is a CoreData error, an alert is shown to the user.

Now that the data comes from CoreData, the ‘Job’ file was deleted. This file was only used for dummy data while building the UI.

Added search functionality by company name:
    - the app gets the search query from the search bar.
    - it performs a fetch request using the query string.
    - it assigns the results to jobsArray and displays the results in the collection view.
    - while the search results are shown, a button opposite the ‘Applications’ label is visible and can be tapped to allow users to go back to the main list.
    - this button is only visible to the user when search results are displayed in the collection view.
    - if the search query returns no results, an alert notifies the user.