Changes made post-PR review.
Removed empty methods from AppDelegate and SceneDelegate.

Removed line spaces between the IBOutlets.

Removed the isTextEntered variable from HomeViewController and rewrote the guard statement in makeSearchQuery() to remove the forcibly unwrapped optionals.

Added the missing ‘private’ keyword to the configure() method in the PopUpButton file.

Removed ‘AnyObject’ from AddNewJobViewControllerDelegate and JobDetailViewControllerDelegate.

Removed the empty init() method in the CoreDataManager file.

Re-indented code in the three view controllers.

Updated the constraints on the storyboard for the ‘Dashboard’ and ‘Applications’ labels, and ‘ReturnToAllButton’. The previous constraints cut off some of the text in the labels on iPhone SE. Now the entire text of both labels is visible on iPhone SE.