Fixed the bug in editing a job from the search results.
Previously a reference was kept to the index of the job application in the array. This was used when editing and deleting job applications. However, when editing a job application from the search results, other applications in the array were affected.

So, this has been replaced with a UUID. UUID attribute was added to the Job entity in the CoreData model. This attribute is now used to edit and delete job applications.